requirements.txt
about-time==4.2.1
alive-progress==3.1.5
atpbar==2.0.3
beautifulsoup4==4.12.3
certifi==2024.7.4
cffi==1.16.0
cftime==1.6.4
charset-normalizer==3.3.2
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.2.1
cycler==0.12.1
DateTime==5.5
deepdiff==7.0.1
dill==0.3.8
fonttools==4.53.1
grapheme==0.6.0
humanfriendly==10.0
idna==3.7
jedi==0.19.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
matplotlib==3.9.1
mdurl==0.1.2
multiprocess==0.70.16
narwhals==1.33.0
netCDF4==1.7.1.post1
numpy==2.0.0
ordered-set==4.1.0
p_tqdm==1.4.2
packaging==24.1
pandas==2.2.2
parallelbar==2.4
parso==0.8.4
pathos==0.3.2
pillow==10.4.0
pip==24.0
plotly==6.0.1
pox==0.3.4
ppft==1.7.6.8
progressbar2==4.4.2
pudb==2022.1.3
pycparser==2.22
pyfar==0.6.8
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-sofa==0.2.0
python-utils==3.8.2
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
scipy==1.14.0
setproctitle==1.3.5
setuptools==71.0.4
six==1.16.0
sofar==1.1.4
sounddevice==0.4.6
soundfile==0.12.1
soupsieve==2.5
tabulate==0.9.0
tqdm==4.66.5
tqdm-multiprocess==0.0.11
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.2.2
urwid==2.6.15
urwid_readline==0.14
wcwidth==0.2.13
wget==3.2
wheel==0.43.0
zope.interface==6.4.post2
